+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v5.5.0 (unreleased)
+
+#### Notes
+Extended support to Image Streamer Rest API version 500((ImageStreamer v3.10)) and API version 600 (ImageStreamer v4.0).
+
+#### Features supported
+Extended support to Image Streamer Rest API version 500 and 600 to the already existing features:
+   - Golden Image
+
 ## v5.4.0
 
 #### Notes